\name{ldahp-package}
\alias{ldahp-package}
\alias{ldahp}
\docType{package}
\title{
Estimation of Latent Dirichlet Allocation Hyperparameters 
}
\description{
This packages is intended to estimate 
the Latent Dirichlet Allocation model hyperparameters 
based on the ratio of LDA model Bayes factors. 
This includes the LDA Gibbs samplers which is 
written in C++ to speedup execution.       
}
\details{
\tabular{ll}{
Package: \tab ldahp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-23\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Authors: Clint P. George and Hani Doss 

Maintainer: Clint P. George <clintpg@ufl.edu>
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
}
\keyword{ lda, Bayes factors, empirical Bayes}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{ }


